requirements.txt
avro
dicttoxml
msgpack
protobuf
PyYAML
xmltodict